+++ src/user_experience/system_tray/system_tray.c
@@ -35,7 +35,7 @@ static bool download_image(const char *url, struct curl_memory_buffer *buffer) {
     }
 
     // Enforce TLS 1.2 or higher for security
-    curl_easy_setopt(curl, CURLOPT_SSLVERSION, CURL_SSLVERSION_TLSv1_2);
+    curl_easy_setopt(curl, CURLOPT_SSLVERSION, (long)CURL_SSLVERSION_TLSv1_2);
 
     curl_memory_buffer_init(buffer);
